♻️ refactor: move activity answer files and more
