+++ api/src/school/application/activity-answer-controller.js
@@ -1,9 +1,9 @@
-import * as activityAnswerSerializer from '../../infrastructure/serializers/jsonapi/activity-answer-serializer.js';
-import { usecases } from '../../domain/usecases/index.js';
+import * as activityAnswerSerializer from '../infrastructure/serializers/activity-answer-serializer.js';
+import { sharedUsecases } from '../../shared/domain/usecases/index.js';
 
 const save = async function (request, h, dependencies = { activityAnswerSerializer }) {
   const { activityAnswer, assessmentId } = dependencies.activityAnswerSerializer.deserialize(request.payload);
-  const createdAnswer = await usecases.correctAnswer({ activityAnswer, assessmentId });
+  const createdAnswer = await sharedUsecases.correctAnswer({ activityAnswer, assessmentId });
 
   return h.response(dependencies.activityAnswerSerializer.serialize(createdAnswer)).created();
 };
