+++ api/src/school/infrastructure/serializers/activity-answer-serializer.js
