+++ api/src/shared/application/assessments/assessment-controller.js
@@ -16,7 +16,7 @@ import {
 } from '../../../../lib/infrastructure/utils/request-response-utils.js';
 import * as certificationChallengeRepository from '../../../certification/shared/infrastructure/repositories/certification-challenge-repository.js';
 
-import { Examiner } from '../../../../lib/domain/models/Examiner.js';
+import { Examiner } from '../../domain/models/Examiner.js';
 import { ValidatorAlwaysOK } from '../../../../lib/domain/models/ValidatorAlwaysOK.js';
 
 const save = async function (request, h, dependencies = { assessmentRepository }) {
