+++ api/src/shared/domain/usecases/index.js
@@ -3,6 +3,9 @@ import { dirname, join } from 'node:path';
 
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
+import * as activityAnswerRepository from '../../../school/infrastructure/repositories/activity-answer-repository.js';
+import * as activityRepository from '../../../school/infrastructure/repositories/activity-repository.js';
+import * as challengeRepository from '../../infrastructure/repositories/challenge-repository.js';
 import * as badgeRepository from '../../infrastructure/repositories/badge-repository.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
@@ -12,6 +15,9 @@ const usecasesWithoutInjectedDependencies = {
 };
 
 const dependencies = {
+  activityAnswerRepository,
+  activityRepository,
+  challengeRepository,
   badgeRepository,
 };
