+++ api/tests/school/unit/infrastructure/serializers/activity-answer-serializer_test.js
@@ -1,6 +1,6 @@
 import { domainBuilder, expect } from '../../../../test-helper.js';
 import { AnswerStatus } from '../../../../../src/shared/domain/models/AnswerStatus.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/activity-answer-serializer.js';
+import * as serializer from '../../../../../src/school/infrastructure/serializers/activity-answer-serializer.js';
 import { ActivityAnswer } from '../../../../../src/school/domain/models/ActivityAnswer.js';
 
 describe('Unit | Serializer | JSONAPI | activity-answer-serializer', function () {
